backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
mysqlclient==2.2.8
sqlparse==0.5.5
tzdata==2025.3
pandas
scikit-learn
joblib